% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_assigning.R
\name{last_observed_stage_table}
\alias{last_observed_stage_table}
\title{Generated data table for observed stages}
\usage{
last_observed_stage_table(data, last_observed_stage)
}
\arguments{
\item{data}{A data frame containing a column with the last observed stages of each row, such as those generated by stage_assigning()}

\item{last_observed_stage}{The name of the column specifying the last observed stage within each row.}
}
\value{
A data table containing information about the distribution of individuals across stages, including the number and percentage of individuals for which it was the final developmental stage and the number and percentage of individuals remaining at each stage.
}
\description{
Generated data table for observed stages
}
\examples{
example_stages = c("Stage1", "Stage2", "Stage3", "Stage4", "Stage5")
assigned_data = dplyr::tribble(
  ~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5, ~last_observed_stage,
  1, "01/01/2000", NA, NA, NA, NA, "Stage1",
  2, "01/01/2000", "02/01/2000", NA, NA, NA, "Stage2",
  3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA, "Stage3",
  4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA, "Stage4",
  5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000", "Stage5",
  6, NA, NA, NA, NA, NA, "no_stage_found"
)
last_observed_stage_table(assigned_data, "last_observed_stage")
}
