% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_assigning.R
\name{snapshotr}
\alias{snapshotr}
\title{Take a snapshot of the stage distribution at a set point in time}
\usage{
snapshotr(data, columns, date, format = "\%d/\%m/\%Y")
}
\arguments{
\item{data}{The data frame being analysed}

\item{columns}{The columns containing dates of stages}

\item{date}{The specific date to test}

\item{format}{The format of the date}
}
\value{
A data frame containing the distributions of each stage
}
\description{
Take a snapshot of the stage distribution at a set point in time
}
\examples{
example_stages = c("Stage1", "Stage2", "Stage3", "Stage4", "Stage5")
example_data = dplyr::tribble(
~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5,
  1, "01/01/2000", NA, NA, NA, NA,
  2, "01/01/2000", "02/01/2000", NA, NA, NA,
  3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA,
  4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA,
  5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000",
  6, NA, NA, NA, NA, NA,
)

snapshotr(example_data, example_stages, "03/01/2000", "\%d/\%m/\%Y")
}
