% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_assigning.R
\name{stage_assigning}
\alias{stage_assigning}
\title{Finds the latest recorded developmental stage}
\usage{
stage_assigning(columns, data)
}
\arguments{
\item{columns}{A character vector specifying the column names of each stage in order.}

\item{data}{The data frame containing information of date an individual reached a specific developmental stage with one row for each individual.}
}
\value{
A new column "last_observed_stage" which contains the last stage found.
}
\description{
Finds the latest recorded developmental stage
}
\examples{
example_stages = c("Stage1", "Stage2", "Stage3", "Stage4", "Stage5")
example_data = dplyr::tribble(
~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5,
  1, "01/01/2000", NA, NA, NA, NA,
  2, "01/01/2000", "02/01/2000", NA, NA, NA,
  3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA,
  4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA,
  5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000",
  6, NA, NA, NA, NA, NA,
)
stage_assigning(columns = example_stages, data = example_data)
}
