% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spanr.R
\name{time_formatr}
\alias{time_formatr}
\title{Converts the type of columns to dates}
\usage{
time_formatr(data, stages, format)
}
\arguments{
\item{data}{The data frame containing the data.}

\item{stages}{The columns containing the stage data.}

\item{format}{The format of the dates in the data frame.}
}
\value{
A data frame with formatted dates.
}
\description{
Converts the type of columns to dates
}
\examples{
example_stages = c("Stage1", "Stage2", "Stage3", "Stage4", "Stage5")
example_data = dplyr::tribble(
  ~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5,
  1, "01/01/2000", NA, NA, NA, NA,
  2, "01/01/2000", "02/01/2000", NA, NA, NA,
  3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA,
  4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA,
  5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000",
  6, NA, NA, NA, NA, NA,
)
time_formatr(example_data, example_stages, "\%d/\%m/\%Y")
}
