% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spanr.R
\name{time_spanr}
\alias{time_spanr}
\title{Generated time span data between developmental stages}
\usage{
time_spanr(data, stages)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{stages}{List of stages containing rows of formatted dates.}
}
\value{
Columns of time differences between sequential stages
}
\description{
Generated time span data between developmental stages
}
\examples{
example_stages = c("Stage1", "Stage2", "Stage3", "Stage4", "Stage5")
example_data = dplyr::tribble(
  ~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5,
  1, "01/01/2000", NA, NA, NA, NA,
  2, "01/01/2000", "02/01/2000", NA, NA, NA,
  3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA,
  4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA,
  5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000",
  6, NA, NA, NA, NA, NA,
)
formatted_data = time_formatr(example_data, example_stages, "\%d/\%m/\%Y")
time_spanr(formatted_data, example_stages)
}
