% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_assigning.R
\name{visualising_distribution}
\alias{visualising_distribution}
\title{Final observed stage distribution graph generation}
\usage{
visualising_distribution(data, stage, factor = FALSE)
}
\arguments{
\item{data}{Data frame containing a column with the last observed stage such as those produced by stage_assigning()}

\item{stage}{The name of the column containing the final observed stage}

\item{factor}{(Optional) Set factor variables, each unique variable will generate a separate bar}
}
\value{
ggplot stacked bar graph showing the distributions of the final observed stage
}
\description{
Final observed stage distribution graph generation
}
\examples{
assigned_data = dplyr::tribble(
  ~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5, ~last_observed_stage,
  1, "01/01/2000", NA, NA, NA, NA, "Stage1",
  2, "01/01/2000", "02/01/2000", NA, NA, NA, "Stage2",
  3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA, "Stage3",
  4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA, "Stage4",
  5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000", "Stage5",
  6, NA, NA, NA, NA, NA, "no_stage_found"
)
visualising_distribution(assigned_data, last_observed_stage)

data_with_factor = dplyr::tribble(
 ~ID, ~ Stage1, ~Stage2, ~Stage3, ~Stage4, ~Stage5, ~last_observed_stage, ~AB,
 1, "01/01/2000", NA, NA, NA, NA, "Stage1", "A",
 2, "01/01/2000", "02/01/2000", NA, NA, NA, "Stage2", "A",
 3, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA, "Stage3", "A",
 4, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA, "Stage4", "A",
 5, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000", "Stage5", "A",
 6, NA, NA, NA, NA, NA, "no_stage_found", "A",
 7, "01/01/2000", NA, NA, NA, NA, "Stage1", "B",
 8, "01/01/2000", "02/01/2000", NA, NA, NA, "Stage2", "B",
 9, "01/01/2000", "02/01/2000", "03/01/2000", NA, NA, "Stage3", "B",
 10, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", NA, "Stage4", "B",
 11, "01/01/2000", "02/01/2000", "03/01/2000", "04/01/2000", "05/01/2000", "Stage5", "B",
 12, NA, NA, NA, NA, NA, "no_stage_found", "B"
)
visualising_distribution(data_with_factor, last_observed_stage, AB)
}
