% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_assigning.R
\name{visualising_survival}
\alias{visualising_survival}
\title{Generates a bar graph containing information about progression through each developmental stage}
\usage{
visualising_survival(
  data,
  stages,
  remaining_percentage,
  remaining_number = NULL
)
}
\arguments{
\item{data}{Data frame containing information about survival at each stage, such as those generated by last_observed_stage_table()}

\item{stages}{Column specifying the name of each stage}

\item{remaining_percentage}{Column specifying the percentage remaining at each developmental stage}

\item{remaining_number}{Column representing the number of individuals remaining at each stage (optional)}
}
\value{
ggplot bar graph with information about percentage survival (and N numbers if provided) at each stage.
}
\description{
Generates a bar graph containing information about progression through each developmental stage
}
\examples{
stage_table = dplyr::tribble(
  ~last_observed_stage, ~n, ~percentage, ~cumulative, ~remaining_n, ~remaining_percentage,
  "Stage1", 1, (1/5)*100, 1, 5, 100,
  "Stage2", 1, (1/5)*100, 2, 4, 80,
  "Stage3", 1, (1/5)*100, 3, 3, 60,
  "Stage4", 1, (1/5)*100, 4, 2, 40,
  "Stage5", 1, (1/5)*100, 5, 1, 20
)
visualising_survival(stage_table, last_observed_stage, remaining_percentage, remaining_n)
visualising_survival(stage_table, last_observed_stage, remaining_percentage)
}
